reorganize supporting documents sections
