+++ src/DashboardUI/src/pages/application/components/application-document.scss
